Separate idempotency tests
