+++ test/Tethos.Moq.Tests/AutoMockingTestTests.cs
@@ -34,20 +34,6 @@
             this.AutoResolver.Should().BeOfType<AutoMoqResolver>();
         }
 
-        [Fact]
-        [Trait("Category", "Integration")]
-        public void Test_Idempotency_ShouldMatchMocks()
-        {
-            // Arrange
-            var expected = this.Container.Resolve<Mock<IMockable>>();
-
-            // Act
-            var actual = this.Container.Resolve<Mock<IMockable>>();
-
-            // Assert
-            actual.Should().Be(expected);
-        }
-
         [Theory]
         [AutoData]
         [Trait("Category", "Integration")]
